+++ backend/src/__tests__/services/reactionService.test.ts
@@ -1,5 +1,5 @@
 import ReactionService from "../../services/reactionService";
-import ReactionRepository from "../../db/reactionRepository";
+import ReactionRepository from "../../repositories/reactionRepository";
 import { Reaction } from "../../models/reaction";
 import { APIResponse } from "../../types/response";
